Separate UserCreate and UserEdit to 2 file
