+++ admin/src/components/UserCreate.js
@@ -54,37 +54,3 @@ export const UserCreate = (props) => (
     </SimpleForm>
   </Create>
 );
-
-export const UserEdit = (props) => (
-  <Edit {...props}>
-    <SimpleForm>
-      <SimpleForm>
-        <TextInput
-          source="name"
-          label={labelName}
-          initialValue="Name" /*todo inicializálni*/
-        />
-        <TextInput
-          source="email"
-          label={labelEmail}
-          initialValue="test@email.com" /*todo inicializálni*/
-        />
-        <PasswordInput
-          source="passord"
-          label={labelPassword}
-          initialValue="password" /*todo inicializálni*/
-        />
-        <DateInput
-          source="date_of_birth"
-          label={labelDateOfBirth}
-          initialValue={new Date()} /*todo inicializálni*/
-        />
-        <TextInput
-          source="license_category"
-          label={labelLicenseCategory}
-          initialValue="X" /*todo inicializálni*/
-        />
-      </SimpleForm>
-    </SimpleForm>
-  </Edit>
-);
